Use board parameter in checkWin and subfunctions
Previously, checkWin did not update the board it used on
each call. This resulted in incorrect behavior upon
resetting the game, where a player would 'win' after only
a few moves. Including a board parameter fixes this by
ensuring that the most up-to-date board is passed to
checkWin (and subfunctions) each time.