Include board as argument in checkAnyWin calls
Previously, a board was not included as a parameter in
checkAnyWin. When this was changed, the two calls to
the function in minimax were not updated with a second
argument. This led to minimax not returning the correct
move in many scenarios. With this change, minimax works
properly once again.